pyproject.toml
[project]
name = "flight-price-tracker"
version = "0.1.0"
description = "Track flight price changes for a route"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "dlt[filesystem]>=1.21.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "PyYAML>=6.0.1",
  "pyarrow>=14.0.0",
]

[project.scripts]
flight-price-tracker = "flight_price_tracker.cli:main"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flight_price_tracker"]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
  "types-PyYAML",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "D"]
ignore = [
  "D203", # incompatible with D211
  "D213", # incompatible with D212
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
